+++ test/test264.js
@@ -1,0 +1,24 @@
+if(typeof exports === 'object') {
+	var assert = require("assert");
+	var alasql = require('..');
+} else {
+	__dirname = '.';
+};
+
+if(typeof exports == 'object') {
+
+describe('Test 264 SPRINTF() function', function() {
+
+  it('1. Test', function(done) {
+
+//    var res = alasql('SELECT VALUE SPRINTF("%d1.1",123.456)');
+    console.log(res);
+//    assert.deepEqual(res,[ { 'MAX(MAX(a),MIN(a))': 8, 'MIN(MAX(a),MIN(a))': 1 } ]);
+    done();    
+  });
+
+
+});
+
+}
+