+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 0.0.50 "Seoul" (21.04.2015 - ...04.2015)
 * CREATE VERTEX and CREATE EDGE syntax
 * Fixed MIN and MAX functions and aggregators #93
-
+* 
 
 ### 0.0.49 "Beijing" (19.04.2015 - 21.04.2015)
 * CREATE CLASS